refactor(tui): split tasks-panel into focused sibling files
The 1,373-LOC tasks-panel.tsx is broken into a lean orchestrator + six
sibling files under `tasks-panel-internals/`. Approach B was chosen so
the public entry point (`@src/.../tasks-panel.tsx`) stays at the same
path — no import churn across the two src consumers and 11 test files.

Split:
- tasks-panel.tsx (298 LOC) — orchestrator (state, effects, layout)
- internals/format.ts (160 LOC) — pure helpers + display constants
- internals/focus-keys.ts (77 LOC) — focus-key plumbing
- internals/signal-rows.tsx (325 LOC) — SignalLine / CommitSignalLine /
  CompactionMarker / StreamSignalRow / InlineKindsBar + SIGNAL_LABEL_COLOR
- internals/evaluation-row.tsx (97 LOC) — EvaluationLine + criterion fusion
- internals/task-row.tsx (322 LOC) — TaskBlock + OrphanSignals
- internals/task-card-parts.tsx (143 LOC) — RecoveryLine / SubStepLine /
  CriteriaBlock + status presentation maps
- internals/keymap.ts (145 LOC) — useTasksPanelInput hook

SIGNAL_LABEL_COLOR is named-re-exported from the orchestrator so the
help-overlay import continues to resolve. Behaviour is identical; all
48 tasks-panel tests pass; pnpm typecheck / lint / deadcode clean.
